+++ ft_substr.c
@@ -6,11 +6,10 @@
 /*   By: haghbal <haghbal@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/13 18:21:24 by haghbal           #+#    #+#             */
-/*   Updated: 2023/11/23 21:33:47 by haghbal          ###   ########.fr       */
+/*   Updated: 2023/12/11 00:51:14 by haghbal          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
 #include "libft.h"
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
@@ -24,13 +23,13 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 		return (NULL);
 	if (start >= (unsigned int)ft_strlen(s))
 		lenn = 0;
-	else if (lenn > (size_t)ft_strlen(s + start))
+	else if (lenn > ft_strlen(s + start))
 		lenn = ft_strlen(s + start);
 	sub = malloc((lenn + 1) * sizeof(char));
 	if (!sub)
 		return (NULL);
 	i = 0;
-	while (start + i < (size_t)ft_strlen(s) && i < lenn)
+	while (start + i < ft_strlen(s) && i < lenn)
 	{
 		sub[i] = s[start + i];
 		i++;
@@ -38,13 +37,3 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	sub[i] = '\0';
 	return (sub);
 }
-
-// int main()
-// {
-// 	const char *s = "bonjour les amis";
-// 	unsigned int start = 4;
-// 	size_t len = 7;
-// 	char *new_str;
-// 	new_str = ft_substr(s, start, len);
-// 	printf("%s", new_str);
-// }
